// src/scrapeui/viewer/types.ts
export type Tab =
  | "configs"
  | "logs"
  | "har"
  | "users"
  | "groups"
  | "roles"
  | "access"
  | "access_logs"
  | "issues"
  | "snapshot"
  | "last_summary"
  | "spec";